requirements.txt
streamlit==1.41.1
tensorflow==2.18.0
pandas>=1.5.0
pillow>=10.0.0
transformers>=4.47.0
python-dotenv>=0.21.0
openai==0.28